refactor(ui): move StatusBar, SessionController and LanguageContext to dedicated subfolders
