+++ src/ui/screens/LiveTranslationScreen.tsx
@@ -2,11 +2,11 @@ import React, { useEffect, useRef, useState } from 'react';
 import { ScrollView, Text, TouchableOpacity, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import { ChevronDown, CircleX, Settings, SwitchCamera } from 'lucide-react-native';
-import { useLanguage } from '../LanguageContext';
+import { useLanguage } from '../context/LanguageContext';
 
 import { ModalityType, SegmentStatus, type TranscriptSegment } from '../../common/models';
-import { StatusBar } from '../StatusBar';
-import { useSessionController } from '../SessionController';
+import { StatusBar } from '../components/StatusBar';
+import { useSessionController } from '../controller/SessionController';
 
 // ---------------------------------------------------------------------------
 // DEV-only test data
